+++ src/PVOutput.Net/Objects/Modules/Implementations/Insolation.cs
@@ -6,7 +6,7 @@ namespace PVOutput.Net.Objects.Modules.Implementations
 {
     internal class Insolation : IInsolation
     {
-        public DateTime Time { get; set; }
+        public TimeSpan Time { get; set; }
         public int Power { get; set; }
         public int Energy { get; set; }
     }