+++ tests/PVOutput.Net.Tests/Modules/Insolation/InsolationServiceTests.cs
@@ -16,6 +16,19 @@ namespace PVOutput.Net.Tests.Modules.Insolation
     public partial class InsolationServiceTests : BaseRequestsTest
     {
         [Test]
+        public async Task InsolationService_GetForOwnSystem_CallsCorrectUri()
+        {
+            PVOutputClient client = TestUtility.GetMockClient(out MockHttpMessageHandler testProvider);
+
+            testProvider.ExpectUriFromBase(GETINSOLATION_URL)
+                        .RespondPlainText(INSOLATION_RESPONSE_BASIC);
+
+            var response = await client.Insolation.GetInsolationForOwnSystemAsync();
+            testProvider.VerifyNoOutstandingExpectation();
+            AssertStandardResponse(response);
+        }
+
+        [Test]
         public async Task InsolationService_GetForSystem_CallsCorrectUri()
         {
             PVOutputClient client = TestUtility.GetMockClient(out MockHttpMessageHandler testProvider);
@@ -25,6 +38,20 @@ namespace PVOutput.Net.Tests.Modules.Insolation
                         .RespondPlainText(INSOLATION_RESPONSE_BASIC);
 
             var response = await client.Insolation.GetInsolationForSystemAsync(54321);
+            testProvider.VerifyNoOutstandingExpectation();
+            AssertStandardResponse(response);
+        }
+
+        [Test]
+        public async Task InsolationService_GetForLocation_CallsCorrectUri()
+        {
+            PVOutputClient client = TestUtility.GetMockClient(out MockHttpMessageHandler testProvider);
+
+            testProvider.ExpectUriFromBase(GETINSOLATION_URL)
+                        .WithQueryString("ll=-84.623970,42.728820")
+                        .RespondPlainText(INSOLATION_RESPONSE_BASIC);
+
+            var response = await client.Insolation.GetInsolationForLocationAsync(new PVCoordinate(-84.62397m, 42.72882m));
             testProvider.VerifyNoOutstandingExpectation();
             AssertStandardResponse(response);
         }
@@ -44,11 +71,11 @@ namespace PVOutput.Net.Tests.Modules.Insolation
             {
                 Assert.AreEqual(0, firstInsolation.Energy);
                 Assert.AreEqual(0, firstInsolation.Power);
-                Assert.AreEqual(DateTime.Today.Add(new TimeSpan(6, 0, 0)), firstInsolation.Time);
+                Assert.AreEqual(new TimeSpan(6, 0, 0), firstInsolation.Time);
 
                 Assert.AreEqual(30, lastInsolation.Energy);
                 Assert.AreEqual(123, lastInsolation.Power);
-                Assert.AreEqual(DateTime.Today.Add(new TimeSpan(6, 35, 0)), lastInsolation.Time);
+                Assert.AreEqual(new TimeSpan(6, 35, 0), lastInsolation.Time);
             });
         }
     }