+++ tests/PVOutput.Net.Tests/Modules/Status/StatusServiceTests.cs
@@ -8,6 +8,8 @@ using PVOutput.Net.Objects.Factories;
 using PVOutput.Net.Objects;
 using PVOutput.Net.Tests.Utils;
 using RichardSzalay.MockHttp;
+using PVOutput.Net.Builders;
+using System.Collections;
 
 namespace PVOutput.Net.Tests.Modules.Status
 {
@@ -55,8 +57,11 @@ namespace PVOutput.Net.Tests.Modules.Status
             testProvider.VerifyNoOutstandingExpectation();
             AssertStandardResponse(response);
 
-            Assert.AreEqual(new DateTime(2020, 1, 31, 14, 40, 0), response.Value.PeakTime);
-            Assert.AreEqual(new DateTime(2020, 1, 31, 15, 5, 0), response.Value.StandbyPowerTime);
+            Assert.Multiple(() =>
+            {
+                Assert.AreEqual(new TimeSpan(14, 40, 0), response.Value.PeakTime);
+                Assert.AreEqual(new TimeSpan(15, 5, 0), response.Value.StandbyPowerTime);
+            });
         }
 
         [Test]
@@ -139,6 +144,21 @@ namespace PVOutput.Net.Tests.Modules.Status
         }
 
         [Test]
+        public async Task StatusService_AddStatus_CallsCorrectUri()
+        {
+            var status = new StatusPostBuilder<IStatusPost>().SetTimeStamp(new DateTime(2020, 1, 1, 12, 22, 0))
+                    .SetGeneration(11000).SetConsumption(9000).Build();
+
+            PVOutputClient client = TestUtility.GetMockClient(out MockHttpMessageHandler testProvider);
+            testProvider.ExpectUriFromBase(ADDSTATUS_URL)
+                        .WithQueryString("d=20200101&t=12:22&v1=11000&v3=9000&n=0")
+                        .RespondPlainText("");
+
+            await client.Status.AddStatusAsync(status);
+            testProvider.VerifyNoOutstandingExpectation();
+        }
+
+        [Test]
         public void StatusService_AddBatchStatus_WithNullStatuses_Throws()
         {
             PVOutputClient client = TestUtility.GetMockClient(out MockHttpMessageHandler testProvider);
@@ -160,6 +180,20 @@ namespace PVOutput.Net.Tests.Modules.Status
             });
         }
 
+        [Test]
+        public async Task StatusService_AddBatchStatus_CallsCorrectUri()
+        {
+            var batchStatus = new StatusPostBuilder<IBatchStatusPost>().SetTimeStamp(new DateTime(2020, 1, 1, 12, 22, 0))
+                    .SetGeneration(11000).SetConsumption(9000).Build();
+
+            PVOutputClient client = TestUtility.GetMockClient(out MockHttpMessageHandler testProvider);
+            testProvider.ExpectUriFromBase(ADDBATCHSTATUS_URL)
+                        .WithQueryString("n=0&data=20200101,12:22,11000,,9000,,,,,,,,,;")
+                        .RespondPlainText("");
+
+            await client.Status.AddBatchStatusAsync(new[] { batchStatus });
+            testProvider.VerifyNoOutstandingExpectation();
+        }
 
         /*
          * Deserialisation tests below
@@ -254,7 +288,7 @@ namespace PVOutput.Net.Tests.Modules.Status
                 Assert.AreEqual(334, result.EnergyGeneration);
                 Assert.AreEqual(1, result.PowerGeneration);
                 Assert.AreEqual(191, result.PeakPower);
-                Assert.AreEqual(DateTime.Today.AddHours(11), result.PeakTime);
+                Assert.AreEqual(new TimeSpan(11, 0, 0), result.PeakTime);
             });
         }
 
@@ -268,12 +302,12 @@ namespace PVOutput.Net.Tests.Modules.Status
                 Assert.AreEqual(334, result.EnergyGeneration);
                 Assert.AreEqual(2, result.PowerGeneration);
                 Assert.AreEqual(82, result.PeakPower);
-                Assert.AreEqual(DateTime.Today.AddHours(14).AddMinutes(40), result.PeakTime);
+                Assert.AreEqual(new TimeSpan(14, 40, 0), result.PeakTime);
 
                 Assert.AreEqual(5811, result.EnergyConsumption);
                 Assert.AreEqual(417, result.PowerConsumption);
                 Assert.AreEqual(255, result.StandbyPower);
-                Assert.AreEqual(DateTime.Today.AddHours(15).AddMinutes(5), result.StandbyPowerTime);
+                Assert.AreEqual(new TimeSpan(15, 5, 0), result.StandbyPowerTime);
             });
         }
 
@@ -287,12 +321,12 @@ namespace PVOutput.Net.Tests.Modules.Status
                 Assert.AreEqual(35302, result.EnergyGeneration);
                 Assert.AreEqual(3, result.PowerGeneration);
                 Assert.AreEqual(5369, result.PeakPower);
-                Assert.AreEqual(DateTime.Today.AddHours(12).AddMinutes(45), result.PeakTime);
+                Assert.AreEqual(new TimeSpan(12, 45, 0), result.PeakTime);
 
                 Assert.AreEqual(31476, result.EnergyConsumption);
                 Assert.AreEqual(606, result.PowerConsumption);
                 Assert.AreEqual(495, result.StandbyPower);
-                Assert.AreEqual(DateTime.Today.AddHours(9).AddMinutes(35), result.StandbyPowerTime);
+                Assert.AreEqual(new TimeSpan(9, 35, 0), result.StandbyPowerTime);
 
                 Assert.AreEqual(18.1d, result.MinimumTemperature);
                 Assert.AreEqual(26.6d, result.MaximumTemperature);