+++ CHANGELOG.md
@@ -12,9 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed bugs with encoding of text parameters `Comments` and `TextMessage` in `OutputPost` and `StatusPost` requests respectively. [#29](https://github.com/pyrocumulus/pvoutput.net/issues/29)
 - Fixed bug with `BatchOutputPostBuilder` accepting an output without both `EnergyGenerated` as wel as `EnergyUsed`. [#29](https://github.com/pyrocumulus/pvoutput.net/issues/29)
 - Breaking - renamed multiple operations on the new `OutputPostBuilders` to accurately map to the corresponding property on IOutputPost etc. [#29](https://github.com/pyrocumulus/pvoutput.net/issues/29)
-- Added structured search methods to the SearchService [#27](https://github.com/pyrocumulus/pvoutput.net/issues/27)
-- Fixed a bug with the SearchService not properly encoding the query text [#27](https://github.com/pyrocumulus/pvoutput.net/issues/27)
-
+- Added structured search methods to the SearchService. [#27](https://github.com/pyrocumulus/pvoutput.net/issues/27)
+- Fixed a bug with the SearchService not properly encoding the query text. [#27](https://github.com/pyrocumulus/pvoutput.net/issues/27)
+- Breaking - all properties respresenting a time are now respresented as `TimeSpan` objects instead of `DateTime`. [#30](https://github.com/pyrocumulus/pvoutput.net/issues/30)
+- Breaking - `PVCoordinate` now uses the `decimal` type for storing the `Latitude` and `Longitude` properties, to avoid weird approximation issues. [#30](https://github.com/pyrocumulus/pvoutput.net/issues/30)
 
 ## [0.6.0] - 2020-03-28